day01/requirements.txt
murf
flask
pyaudio
requests
python-dotenv
